refctor(sequencing metrics parser) (#3241) (patch)
# Description

Refactoring of sequencing metrics parser
